set default minsize = 1
why?  because there are lots of empty files, and a lot of times the
mtimes are actually relevant - sometimes you touch a file for the
specific purpose of marking a time.

the mtimes of course get blown away for all but the oldest if you are
hardlinking, which makes sense for pretty much everything except empty
files.