avoid unnecessary stat calls
The stat calls are only needed for comparing files that would get
linked; that is, files with the same shasum.  st_mtime is used for
finding the oldest in a set of identical files, and st_ino is used to
avoid re-linking already hardlinked files.  In any case, stat info is
never needed for a file that does not share a shasum with any other
file.

Introduce a LazyStat descriptor for an FInfo class which sorts first on
sha and then on st.st_mtime, where the stat happens only the first time
when the fi.st member gets accessed.  Fun!

Why bother?  Stat is slow on some filesystems, like ntfs-3g and afs.